pyproject.toml
[tool.poetry]
name = "gotravel-backend"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.105.0"
pydantic = "^2.5.2"
sqlalchemy = "^2.0.23"
expiringdictx = "^1.0.1"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
alembic = "^1.13.0"
uvicorn = "^0.24.0.post1"
python-multipart = "^0.0.6"

[tool.poetry.group.lint.dependencies]
black = "^23.7.0"
ipdb = "^0.13.4"
isort = "^5.10.1"
nonemoji = "^0.1.4"
pre-commit = "^3.3.0"
ruff = ">=0.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
select = ["E", "W", "F", "UP", "C", "T", "PYI", "PT", "Q"]
ignore = ["E402", "C901", "PT023"]

line-length = 120
target-version = "py310"

[tool.black]
line-length = 120
preview = true
target-version = ["py310", "py311"]
include = '\.pyi?$'
extend-exclude = '''
'''

[tool.isort]
profile = "black"
line_length = 120
skip_gitignore = true
length_sort = true
force_sort_within_sections = true
src_paths = ["backend"]
extra_standard_library = ["typing_extensions"]
